financial-news-streamlit_2/requirements.txt
streamlit==1.28.1
streamlit-option-menu==0.3.6
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.32
feedparser==6.0.10
openai==1.3.0
google-generativeai==0.3.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.5
scikit-learn==1.3.2
